save and

+++ App.Game/Controllers/GameModelsController.cs
@@ -15,7 +15,7 @@ namespace App.Game.Controllers
 {
     public class GameModelsController : Controller
     {
-        private ApplicationDbContext db = new ApplicationDbContext();
+       private ApplicationDbContext db = new ApplicationDbContext();
        private ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(System.Web.HttpContext.Current.User.Identity.GetUserId());
 
         // GET: GameModels
@@ -23,19 +23,9 @@ namespace App.Game.Controllers
         {
             PessoaModel pessoa = db.Pessoa.Where(d=> d.ApplicationUserID == user.Id).First();
 
-            //PessoaModel pessoa1 = db.Pessoa.Where(d => d.Id == 2).First();
-
-            //var teste = from d in db.PessoaGame from p in db.Game where d.pessoa_pessoa_id == pessoa.Id select d; 
-            //ICollection<GameModel> lista = db.Game.Where(t => t.PessoaGame.pessoa_pessoa_id == pessoa.Id).ToList();
             ViewBag.Title = "Inicial";
-
-           // EmprestimoModel emp = new EmprestimoModel();
-           // PessoaModel pessoa3 = new PessoaModel();
-           // pessoa.PessoaMe.Add(pessoa1);
-           //// pessoa3.PessoaFrinds.Add(pessoa1);
-           // pessoa.PessoaFrinds.Add(pessoa);
-            // db.Pessoa.Add(pessoa3);
             IEnumerable<GameModel> c = db.PessoaGame.Where(e => e.pessoa_pessoa_id == pessoa.Id).Select(e => e.Game);
+
 
             return View(c.ToList());
         }
@@ -72,19 +62,18 @@ namespace App.Game.Controllers
             {
                 PessoaModel pessoa = db.Pessoa.Where(c => c.ApplicationUserID == user.Id).First();
                 db.Game.Add(gameModel);
-                PessoaGameModel pessoaGame = new PessoaGameModel();
-                //db.SaveChangesAsync();
-                pessoaGame.Pessoa = pessoa;
-                pessoaGame.Game = gameModel;
-                pessoaGame.game_game_id = gameModel.GameId;
-                pessoaGame.pessoa_pessoa_id = pessoa.Id;
+                PessoaGameModel pessoaGame = new PessoaGameModel
+                {
+                    Pessoa = pessoa,
+                    Game = gameModel,
+                    game_game_id = gameModel.GameId,
+                    pessoa_pessoa_id = pessoa.Id
+                };
                 db.PessoaGame.Add(pessoaGame);
                 db.SaveChanges();
 
-
-
-
-                var teste = gameModel.GameId;
+                ViewBag.classe = "success";
+                ViewBag.msg = "Jogo cadastrado com sucesso!";
 
                 return RedirectToAction("Index");
             }
@@ -143,9 +132,25 @@ namespace App.Game.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+
             GameModel gameModel = db.Game.Find(id);
+            
+            
+            var emprestado = (from p in db.Emprestismo where p.Game_id == gameModel.GameId select p).Count();
+
+            if (emprestado > 0)
+            {
+                ModelState.AddModelError("Descricao", "Não é Possível Excluir");
+                
+                return View(gameModel);
+            }
+            PessoaGameModel pessoagame = db.PessoaGame.Where(e => e.game_game_id == gameModel.GameId && e.pessoa_pessoa_id == 1).First();
+
+            db.PessoaGame.Remove(pessoagame);
             db.Game.Remove(gameModel);
             db.SaveChanges();
+            this.TempData["classe"] ="success";
+            this.TempData["msg"] = "Jogo excluído com sucesso!";
             return RedirectToAction("Index");
         }
